+++ ktrain/core.py
@@ -132,9 +132,22 @@ class Learner(ABC):
         return wds
 
 
-    def set_weight_decay(self, wd=1e-2):
+    def set_weight_decay(self, wd=0.005):
         """
-        sets global weight decay layer-by-layer
+        Sets global weight decay layer-by-layer using L2 regularization.
+
+        NOTE: Weight decay can be implemented in the form of
+              L2 regularization, which is the case with Keras.
+              Thus, he weight decay value must be divided by
+              2 to obtain similar behavior.
+              The default weight decay here is 0.01/2 = 0.005.
+              See here for more information: 
+              https://bbabenko.github.io/weight-decay/
+        Args:
+          wd(float): weight decay (see note above)
+        Returns:
+          None
+              
         """
         for layer in self.model.layers:
             if hasattr(layer, 'kernel_regularizer'):
@@ -400,8 +413,7 @@ class Learner(ABC):
 
 
 
-    def lr_find(self, epochs=5, start_lr=1e-7, 
-                verbose=1):
+    def lr_find(self, start_lr=1e-7, epochs=None, verbose=1):
         """
         Plots loss as learning rate is increased.
         Highest learning rate corresponding to a still
@@ -411,6 +423,7 @@ class Learner(ABC):
 
         Args:
             epochs (int): maximum number of epochs to simulate training
+                          If None, chosen automatically.
             start_lr (float): smallest lr to start simulation
             verbose (bool): specifies how much output to print
         Returns:
@@ -430,7 +443,8 @@ class Learner(ABC):
         try:
             # track and plot learning rates
             self.lr_finder = LRFinder(self.model)
-            self.lr_finder.find(self.train_data, start_lr=start_lr, end_lr=10, epochs=5,
+            self.lr_finder.find(self.train_data, start_lr=start_lr, end_lr=10, 
+                                epochs=epochs,
                                 workers=self.workers, 
                                 use_multiprocessing=self.use_multiprocessing, 
                                 verbose=verbose)