+++ ktrain/lroptimize/lrfinder.py
@@ -1,10 +1,12 @@
 from matplotlib import pyplot as plt
 import numpy as np
 import tempfile
+import math
 from keras.callbacks import LambdaCallback, Callback
 import keras.backend as K
 from .. import utils as U
 
+SAMPLE_SIZE = 2048.
 
 class LRFinder:
     """
@@ -51,7 +53,7 @@ class LRFinder:
         K.set_value(self.model.optimizer.lr, lr)
 
 
-    def find(self, train_data, start_lr, end_lr, epochs=1, batch_size=U.DEFAULT_BS,
+    def find(self, train_data, start_lr, end_lr, epochs=None, batch_size=U.DEFAULT_BS,
              workers=1, use_multiprocessing=False, verbose=1):
         """
         Track loss as learning rate is increased.
@@ -73,6 +75,10 @@ class LRFinder:
         else:
             use_gen = False
             steps_per_epoch = np.ceil(num_samples/batch_size)
+
+        if epochs is None:
+            epochs = math.ceil(SAMPLE_SIZE/steps_per_epoch)
+
 
         num_batches = epochs * steps_per_epoch
         self.lr_mult = (end_lr / start_lr) ** (1 / num_batches)