+++ CHANGELOG.md
@@ -6,6 +6,17 @@ Most recent releases are shown at the top. Each release shows:
 - **Changed**: Additional parameters, changes to inputs or outputs, etc
 - **Fixed**: Bug fixes that don't change documented behaviour
 
+## 0.1.7 (2019-05-24)
+
+### New:
+- BIGRU text classifier: Bidirectional GRU using pretrained word embeddings
+
+### Changed:
+- Epochs are calculated automatically in ```LRFinder```
+
+### Fixed:
+- Number of epochs that ```Learner.lr_find``` runs can be explicitly set again
+
 
 ## 0.1.6 (2019-05-03)
 
@@ -25,7 +36,6 @@ Most recent releases are shown at the top. Each release shows:
 - **cycle\_momentum** argument for both```autofit``` and ```fit_onecycle``` method that will cycle momentum between 0.95 and 0.85 as described in [this paper](https://arxiv.org/abs/1803.09820)
 - ```Learner.plot``` method that will plot training-validation loss, LR schedule, or momentum schedule
 - added ```set_weight_decay``` and ```get_weight_decay``` methods to get/set "global" weight decay in Keras
-- 
 
 ### Changed:
 - ```vision.data.preview_data_aug``` now displays images in rows by default